backend/requirements.txt
python==3.11
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
langchain==0.0.350
langchain-groq
groq==0.4.2
requests==2.31.0